auth_service/requirements.txt
flask
requests
python-dotenv
flask-sqlalchemy
pyjwt